Reworked tests to use actor features
